refactor kmt invoke task
